refactor(decks): load create page cards asynchronously
